incorporated and edited logistic infos that Samuele forwarded me from Jacob.  Next time, please just check things in yourself, there is no need to route this through my mailbox.
I think we should try to send out this announcement real soon
now, it has been dragging on some weeks to long, if you ask me,
considering that we knew the time and basic place of the sprint
since early May if not earlier.

So please review the announcement and correct/fix/comment it
so we can send things out on the weekend.