a first draft of the 22C3 proposal, everybody who wants to participate: coordinate with me and Carl and most important: add your speaker information!
comments on the talk scope welcome.  I know
that it's pretty stuffed but i also think
that it doesn't make sense (as opposed to
EuroPython conferences) to go into too many
details. Rather i'd like to touch areas
and allow people to have questions.
(The talk is 60 minutes overall, but
45 minutes and then questions is recommended)